# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dog",
        "description": "Documentação da API Dog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/posts/upload": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Upload de post",
                "description": "Upload a user photo.",
                "operationId": "94e6aa000963d65bcda17584217db819",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo",
                                    "title"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "Photo file (jpeg, png, jpg, gif, webp, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "weight": {
                                        "description": "Weight (optional)",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "age": {
                                        "description": "Age (optional)",
                                        "type": "integer"
                                    },
                                    "title": {
                                        "description": "Title",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "Description (optional)",
                                        "type": "string",
                                        "maxLength": 2000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post uploaded successfully."
                                        },
                                        "post": {
                                            "properties": {
                                                "original_path": {
                                                    "type": "string"
                                                },
                                                "feed_path": {
                                                    "type": "string"
                                                },
                                                "thumb_path": {
                                                    "type": "string"
                                                },
                                                "original_url": {
                                                    "type": "string"
                                                },
                                                "feed_url": {
                                                    "type": "string"
                                                },
                                                "thumb_url": {
                                                    "type": "string"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "age": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Buscar um post por ID ou UUID",
                "description": "Buscar um post.",
                "operationId": "19d7b55d7d5efd7bc27ed7c01d74ca95",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID ou UUID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "original_path": {
                                                    "type": "string"
                                                },
                                                "feed_path": {
                                                    "type": "string"
                                                },
                                                "thumb_path": {
                                                    "type": "string"
                                                },
                                                "original_url": {
                                                    "type": "string"
                                                },
                                                "feed_url": {
                                                    "type": "string"
                                                },
                                                "thumb_url": {
                                                    "type": "string"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "age": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "is_published": {
                                                    "type": "boolean"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Atualizar metadados de um post",
                "description": "Atualizar metadados de um post.",
                "operationId": "97ae61bfc5c99543e90d6f4136332c12",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID or UUID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "weight": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 2000
                                    },
                                    "is_published": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "photo": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Deletar um post",
                "description": "Delete a photo and its stored files.",
                "operationId": "f5991cc3c6e9a5fcb5b97135fbe27b0c",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID or UUID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Photo deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Photo not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}